+++ tardis/montecarlo/src/cmontecarlo.c
@@ -1,4 +1,5 @@
 #include "cmontecarlo.h"
+#include ""
 
 rk_state mt_state;
 
@@ -6,245 +7,6 @@ void
 initialize_random_kit (unsigned long seed)
 {
   rk_seed (seed, &mt_state);
-}
-
-INLINE tardis_error_t
-line_search (double *nu, double nu_insert, int64_t number_of_lines,
-	     int64_t * result)
-{
-  tardis_error_t ret_val = TARDIS_ERROR_OK;
-  int64_t imin, imax;
-  imin = 0;
-  imax = number_of_lines - 1;
-  if (nu_insert > nu[imin])
-    {
-      *result = imin;
-    }
-  else if (nu_insert < nu[imax])
-    {
-      *result = imax + 1;
-    }
-  else
-    {
-      ret_val = reverse_binary_search (nu, nu_insert, imin, imax, result);
-      *result = *result + 1;
-    }
-  return ret_val;
-}
-
-inline tardis_error_t
-reverse_binary_search (double *x, double x_insert,
-		       int64_t imin, int64_t imax, int64_t * result)
-{
-  /*
-     Have in mind that *x points to a reverse sorted array.
-     That is large values will have small indices and small ones
-     will have large indices.
-   */
-  tardis_error_t ret_val = TARDIS_ERROR_OK;
-  if (x_insert > x[imin] || x_insert < x[imax])
-    {
-      ret_val = TARDIS_ERROR_BOUNDS_ERROR;
-    }
-  else
-    {
-      int imid = (imin + imax) / 2;
-      while (imax - imin > 2)
-	{
-	  if (x[imid] < x_insert)
-	    {
-	      imax = imid + 1;
-	    }
-	  else
-	    {
-	      imin = imid;
-	    }
-	  imid = (imin + imax) / 2;
-	}
-      if (imax - imin == 2 && x_insert < x[imin + 1])
-	{
-	  *result = imin + 1;
-	}
-      else
-	{
-	  *result = imin;
-	}
-    }
-  return ret_val;
-}
-
-inline tardis_error_t
-binary_search (double *x, double x_insert, int64_t imin,
-	       int64_t imax, int64_t * result)
-{
-  /*
-     Have in mind that *x points to a sorted array.
-     Like [1,2,3,4,5,...]
-   */
-  int imid;
-  tardis_error_t ret_val = TARDIS_ERROR_OK;
-  if (x_insert < x[imin] || x_insert > x[imax])
-    {
-      ret_val = TARDIS_ERROR_BOUNDS_ERROR;
-    }
-  else
-    {
-      while (imax >= imin)
-	{
-	  imid = (imin + imax) / 2;
-	  if (x[imid] == x_insert)
-	    {
-	      *result = imid;
-	      break;
-	    }
-	  else if (x[imid] < x_insert)
-	    {
-	      imin = imid + 1;
-	    }
-	  else
-	    {
-	      imax = imid - 1;
-	    }
-	}
-      if (imax - imid == 2 && x_insert < x[imin + 1])
-	{
-	  *result = imin;
-	}
-      else
-	{
-	  *result = imin;
-	}
-    }
-  return ret_val;
-}
-
-INLINE double
-rpacket_doppler_factor (rpacket_t * packet, storage_model_t * storage)
-{
-  return 1.0 -
-    rpacket_get_mu (packet) * rpacket_get_r (packet) *
-    storage->inverse_time_explosion * INVERSE_C;
-}
-
-INLINE double
-compute_distance2boundary (rpacket_t * packet, storage_model_t * storage)
-{
-  double r = rpacket_get_r (packet);
-  double mu = rpacket_get_mu (packet);
-  double r_outer = storage->r_outer[rpacket_get_current_shell_id (packet)];
-  double r_inner = storage->r_inner[rpacket_get_current_shell_id (packet)];
-  double d_outer =
-    sqrt (r_outer * r_outer + ((mu * mu - 1.0) * r * r)) - (r * mu);
-  double d_inner;
-  double result;
-  if (rpacket_get_recently_crossed_boundary (packet) == 1)
-    {
-      rpacket_set_next_shell_id (packet, 1);
-      return d_outer;
-    }
-  else
-    {
-      double check = r_inner * r_inner + (r * r * (mu * mu - 1.0));
-      if (check < 0.0)
-	{
-	  rpacket_set_next_shell_id (packet, 1);
-	  return d_outer;
-	}
-      else
-	{
-	  d_inner = mu < 0.0 ? -r * mu - sqrt (check) : MISS_DISTANCE;
-	}
-    }
-  if (d_inner < d_outer)
-    {
-      rpacket_set_next_shell_id (packet, -1);
-      return d_inner;
-    }
-  else
-    {
-      rpacket_set_next_shell_id (packet, 1);
-      return d_outer;
-    }
-}
-
-INLINE tardis_error_t
-compute_distance2line (rpacket_t * packet, storage_model_t * storage,
-		       double *result)
-{
-  tardis_error_t ret_val = TARDIS_ERROR_OK;
-  if (rpacket_get_last_line (packet))
-    {
-      *result = MISS_DISTANCE;
-    }
-  else
-    {
-      double r = rpacket_get_r (packet);
-      double mu = rpacket_get_mu (packet);
-      double nu = rpacket_get_nu (packet);
-      double nu_line = rpacket_get_nu_line (packet);
-      double t_exp = storage->time_explosion;
-      double inverse_t_exp = storage->inverse_time_explosion;
-      int64_t cur_zone_id = rpacket_get_current_shell_id (packet);
-      double comov_nu, doppler_factor;
-      doppler_factor = 1.0 - mu * r * inverse_t_exp * INVERSE_C;
-      comov_nu = nu * doppler_factor;
-      if (comov_nu < nu_line)
-	{
-	  if (rpacket_get_next_line_id (packet) == storage->no_of_lines - 1)
-	    {
-	      fprintf (stderr, "last_line = %f\n",
-		       storage->
-		       line_list_nu[rpacket_get_next_line_id (packet) - 1]);
-	      fprintf (stderr, "Last line in line list reached!");
-	    }
-	  else if (rpacket_get_next_line_id (packet) == 0)
-	    {
-	      fprintf (stderr, "First line in line list!");
-	      fprintf (stderr, "next_line = %f\n",
-		       storage->
-		       line_list_nu[rpacket_get_next_line_id (packet) + 1]);
-	    }
-	  else
-	    {
-	      fprintf (stderr, "last_line = %f\n",
-		       storage->
-		       line_list_nu[rpacket_get_next_line_id (packet) - 1]);
-	      fprintf (stderr, "next_line = %f\n",
-		       storage->
-		       line_list_nu[rpacket_get_next_line_id (packet) + 1]);
-	    }
-	  fprintf (stderr, "ERROR: Comoving nu less than nu_line!\n");
-	  fprintf (stderr, "comov_nu = %f\n", comov_nu);
-	  fprintf (stderr, "nu_line = %f\n", nu_line);
-	  fprintf (stderr, "(comov_nu - nu_line) / nu_line = %f\n",
-		   (comov_nu - nu_line) / nu_line);
-	  fprintf (stderr, "r = %f\n", r);
-	  fprintf (stderr, "mu = %f\n", mu);
-	  fprintf (stderr, "nu = %f\n", nu);
-	  fprintf (stderr, "doppler_factor = %f\n", doppler_factor);
-	  fprintf (stderr, "cur_zone_id = %d\n", cur_zone_id);
-	  ret_val = TARDIS_ERROR_COMOV_NU_LESS_THAN_NU_LINE;
-	}
-      else
-	{
-	  *result = ((comov_nu - nu_line) / nu) * C * t_exp;
-	}
-    }
-  return ret_val;
-}
-
-INLINE double
-compute_distance2electron (rpacket_t * packet, storage_model_t * storage)
-{
-  if (rpacket_get_virtual_packet (packet) > 0)
-    {
-      return MISS_DISTANCE;
-    }
-  double inverse_ne =
-    storage->
-    inverse_electron_densities[rpacket_get_current_shell_id (packet)] *
-    storage->inverse_sigma_thomson;
-  return rpacket_get_tau_event (packet) * inverse_ne;
 }
 
 INLINE int64_t
@@ -390,235 +152,6 @@ montecarlo_one_packet (storage_model_t * storage, rpacket_t * packet,
 	}
     }
   return reabsorbed;
-}
-
-void
-move_packet_across_shell_boundary (rpacket_t * packet,
-				   storage_model_t * storage, double distance)
-{
-  double comov_energy, doppler_factor, comov_nu, inverse_doppler_factor;
-  move_packet (packet, storage, distance);
-  if (rpacket_get_virtual_packet (packet) > 0)
-    {
-      double delta_tau_event = distance *
-	storage->electron_densities[rpacket_get_current_shell_id (packet)] *
-	storage->sigma_thomson;
-      rpacket_set_tau_event (packet,
-			     rpacket_get_tau_event (packet) +
-			     delta_tau_event);
-    }
-  else
-    {
-      rpacket_reset_tau_event (packet);
-    }
-  if ((rpacket_get_current_shell_id (packet) < storage->no_of_shells - 1
-       && rpacket_get_next_shell_id (packet) == 1)
-      || (rpacket_get_current_shell_id (packet) > 0
-	  && rpacket_get_next_shell_id (packet) == -1))
-    {
-      rpacket_set_current_shell_id (packet,
-				    rpacket_get_current_shell_id (packet) +
-				    rpacket_get_next_shell_id (packet));
-      rpacket_set_recently_crossed_boundary (packet,
-					     rpacket_get_next_shell_id
-					     (packet));
-    }
-  else if (rpacket_get_next_shell_id (packet) == 1)
-    {
-      rpacket_set_status (packet, TARDIS_PACKET_STATUS_EMITTED);
-    }
-  else if ((storage->reflective_inner_boundary == 0) ||
-	   (rk_double (&mt_state) > storage->inner_boundary_albedo))
-    {
-      rpacket_set_status (packet, TARDIS_PACKET_STATUS_REABSORBED);
-    }
-  else
-    {
-      doppler_factor = rpacket_doppler_factor (packet, storage);
-      comov_nu = rpacket_get_nu (packet) * doppler_factor;
-      comov_energy = rpacket_get_energy (packet) * doppler_factor;
-      rpacket_set_mu (packet, rk_double (&mt_state));
-      inverse_doppler_factor = 1.0 / rpacket_doppler_factor (packet, storage);
-      rpacket_set_nu (packet, comov_nu * inverse_doppler_factor);
-      rpacket_set_energy (packet, comov_energy * inverse_doppler_factor);
-      rpacket_set_recently_crossed_boundary (packet, 1);
-      if (rpacket_get_virtual_packet_flag (packet) > 0)
-	{
-	  montecarlo_one_packet (storage, packet, -2);
-	}
-    }
-}
-
-void
-montecarlo_thomson_scatter (rpacket_t * packet, storage_model_t * storage,
-			    double distance)
-{
-  double comov_energy, doppler_factor, comov_nu, inverse_doppler_factor;
-  doppler_factor = move_packet (packet, storage, distance);
-  comov_nu = rpacket_get_nu (packet) * doppler_factor;
-  comov_energy = rpacket_get_energy (packet) * doppler_factor;
-  rpacket_set_mu (packet, 2.0 * rk_double (&mt_state) - 1.0);
-  inverse_doppler_factor = 1.0 / rpacket_doppler_factor (packet, storage);
-  rpacket_set_nu (packet, comov_nu * inverse_doppler_factor);
-  rpacket_set_energy (packet, comov_energy * inverse_doppler_factor);
-  rpacket_reset_tau_event (packet);
-  rpacket_set_recently_crossed_boundary (packet, 0);
-  storage->last_interaction_type[storage->current_packet_id] = 1;
-  if (rpacket_get_virtual_packet_flag (packet) > 0)
-    {
-      montecarlo_one_packet (storage, packet, 1);
-    }
-}
-
-void
-montecarlo_line_scatter (rpacket_t * packet, storage_model_t * storage,
-			 double distance)
-{
-  double comov_energy = 0.0;
-  int64_t emission_line_id = 0;
-  double old_doppler_factor = 0.0;
-  double inverse_doppler_factor = 0.0;
-  double tau_line = 0.0;
-  double tau_electron = 0.0;
-  double tau_combined = 0.0;
-  bool virtual_close_line = false;
-  int64_t j_blue_idx = -1;
-  if (rpacket_get_virtual_packet (packet) == 0)
-    {
-      j_blue_idx =
-	rpacket_get_current_shell_id (packet) *
-	storage->line_lists_j_blues_nd + rpacket_get_next_line_id (packet);
-      increment_j_blue_estimator (packet, storage, distance, j_blue_idx);
-    }
-  tau_line =
-    storage->line_lists_tau_sobolevs[rpacket_get_current_shell_id (packet) *
-				     storage->line_lists_tau_sobolevs_nd +
-				     rpacket_get_next_line_id (packet)];
-  tau_electron =
-    storage->sigma_thomson *
-    storage->electron_densities[rpacket_get_current_shell_id (packet)] *
-    distance;
-  tau_combined = tau_line + tau_electron;
-  rpacket_set_next_line_id (packet, rpacket_get_next_line_id (packet) + 1);
-  if (rpacket_get_next_line_id (packet) == storage->no_of_lines)
-    {
-      rpacket_set_last_line (packet, true);
-    }
-  if (rpacket_get_virtual_packet (packet) > 0)
-    {
-      rpacket_set_tau_event (packet,
-			     rpacket_get_tau_event (packet) + tau_line);
-    }
-  else if (rpacket_get_tau_event (packet) < tau_combined)
-    {
-      old_doppler_factor = move_packet (packet, storage, distance);
-      rpacket_set_mu (packet, 2.0 * rk_double (&mt_state) - 1.0);
-      inverse_doppler_factor = 1.0 / rpacket_doppler_factor (packet, storage);
-      comov_energy = rpacket_get_energy (packet) * old_doppler_factor;
-      rpacket_set_energy (packet, comov_energy * inverse_doppler_factor);
-      storage->last_line_interaction_in_id[storage->current_packet_id] =
-	rpacket_get_next_line_id (packet) - 1;
-      storage->last_line_interaction_shell_id[storage->current_packet_id] =
-	rpacket_get_current_shell_id (packet);
-      storage->last_interaction_type[storage->current_packet_id] = 2;
-      if (storage->line_interaction_id == 0)
-	{
-	  emission_line_id = rpacket_get_next_line_id (packet) - 1;
-	}
-      else if (storage->line_interaction_id >= 1)
-	{
-	  emission_line_id = macro_atom (packet, storage);
-	}
-      storage->last_line_interaction_out_id[storage->current_packet_id] =
-	emission_line_id;
-      rpacket_set_nu (packet,
-		      storage->line_list_nu[emission_line_id] *
-		      inverse_doppler_factor);
-      rpacket_set_nu_line (packet, storage->line_list_nu[emission_line_id]);
-      rpacket_set_next_line_id (packet, emission_line_id + 1);
-      rpacket_reset_tau_event (packet);
-      rpacket_set_recently_crossed_boundary (packet, 0);
-      if (rpacket_get_virtual_packet_flag (packet) > 0)
-	{
-	  virtual_close_line = false;
-	  if (!rpacket_get_last_line (packet) &&
-	      fabs (storage->line_list_nu[rpacket_get_next_line_id (packet)] -
-		    rpacket_get_nu_line (packet)) /
-	      rpacket_get_nu_line (packet) < 1e-7)
-	    {
-	      virtual_close_line = true;
-	    }
-	  // QUESTIONABLE!!!
-	  bool old_close_line = rpacket_get_close_line (packet);
-	  rpacket_set_close_line (packet, virtual_close_line);
-	  montecarlo_one_packet (storage, packet, 1);
-	  rpacket_set_close_line (packet, old_close_line);
-	  virtual_close_line = false;
-	}
-    }
-  else
-    {
-      rpacket_set_tau_event (packet,
-			     rpacket_get_tau_event (packet) - tau_line);
-    }
-  if (!rpacket_get_last_line (packet) &&
-      fabs (storage->line_list_nu[rpacket_get_next_line_id (packet)] -
-	    rpacket_get_nu_line (packet)) / rpacket_get_nu_line (packet) <
-      1e-7)
-    {
-      rpacket_set_close_line (packet, true);
-    }
-}
-
-INLINE void
-montecarlo_compute_distances (rpacket_t * packet, storage_model_t * storage)
-{
-  // Check if the last line was the same nu as the current line.
-  if (rpacket_get_close_line (packet))
-    {
-      // If so set the distance to the line to 0.0
-      rpacket_set_d_line (packet, 0.0);
-      // Reset close_line.
-      rpacket_set_close_line (packet, false);
-    }
-  else
-    {
-      rpacket_set_d_boundary (packet,
-			      compute_distance2boundary (packet, storage));
-      double d_line;
-      compute_distance2line (packet, storage, &d_line);
-      rpacket_set_d_line (packet, d_line);
-      rpacket_set_d_electron (packet,
-			      compute_distance2electron (packet, storage));
-    }
-}
-
-INLINE montecarlo_event_handler_t
-get_event_handler (rpacket_t * packet, storage_model_t * storage,
-		   double *distance)
-{
-  double d_boundary, d_electron, d_line;
-  montecarlo_compute_distances (packet, storage);
-  d_boundary = rpacket_get_d_boundary (packet);
-  d_electron = rpacket_get_d_electron (packet);
-  d_line = rpacket_get_d_line (packet);
-  montecarlo_event_handler_t handler;
-  if (d_line <= d_boundary && d_line <= d_electron)
-    {
-      *distance = d_line;
-      handler = &montecarlo_line_scatter;
-    }
-  else if (d_boundary <= d_electron)
-    {
-      *distance = d_boundary;
-      handler = &move_packet_across_shell_boundary;
-    }
-  else
-    {
-      *distance = d_electron;
-      handler = &montecarlo_thomson_scatter;
-    }
-  return handler;
 }
 
 int64_t
