+++ tardis/montecarlo/src/geometry.c
@@ -1,0 +1,145 @@
+#include "cmontecarlo.h"
+
+INLINE double
+compute_distance2boundary (rpacket_t * packet, storage_model_t * storage)
+{
+  double r = rpacket_get_r (packet);
+  double mu = rpacket_get_mu (packet);
+  double r_outer = storage->r_outer[rpacket_get_current_shell_id (packet)];
+  double r_inner = storage->r_inner[rpacket_get_current_shell_id (packet)];
+  double d_outer =
+    sqrt (r_outer * r_outer + ((mu * mu - 1.0) * r * r)) - (r * mu);
+  double d_inner;
+  double result;
+  if (rpacket_get_recently_crossed_boundary (packet) == 1)
+    {
+      rpacket_set_next_shell_id (packet, 1);
+      return d_outer;
+    }
+  else
+    {
+      double check = r_inner * r_inner + (r * r * (mu * mu - 1.0));
+      if (check < 0.0)
+	{
+	  rpacket_set_next_shell_id (packet, 1);
+	  return d_outer;
+	}
+      else
+	{
+	  d_inner = mu < 0.0 ? -r * mu - sqrt (check) : MISS_DISTANCE;
+	}
+    }
+  if (d_inner < d_outer)
+    {
+      rpacket_set_next_shell_id (packet, -1);
+      return d_inner;
+    }
+  else
+    {
+      rpacket_set_next_shell_id (packet, 1);
+      return d_outer;
+    }
+}
+
+INLINE tardis_error_t
+compute_distance2line (rpacket_t * packet, storage_model_t * storage,
+		       double *result)
+{
+  tardis_error_t ret_val = TARDIS_ERROR_OK;
+  if (rpacket_get_last_line (packet))
+    {
+      *result = MISS_DISTANCE;
+    }
+  else
+    {
+      double r = rpacket_get_r (packet);
+      double mu = rpacket_get_mu (packet);
+      double nu = rpacket_get_nu (packet);
+      double nu_line = rpacket_get_nu_line (packet);
+      double t_exp = storage->time_explosion;
+      double inverse_t_exp = storage->inverse_time_explosion;
+      int64_t cur_zone_id = rpacket_get_current_shell_id (packet);
+      double comov_nu, doppler_factor;
+      doppler_factor = 1.0 - mu * r * inverse_t_exp * INVERSE_C;
+      comov_nu = nu * doppler_factor;
+      if (comov_nu < nu_line)
+	{
+	  if (rpacket_get_next_line_id (packet) == storage->no_of_lines - 1)
+	    {
+	      fprintf (stderr, "last_line = %f\n",
+		       storage->
+		       line_list_nu[rpacket_get_next_line_id (packet) - 1]);
+	      fprintf (stderr, "Last line in line list reached!");
+	    }
+	  else if (rpacket_get_next_line_id (packet) == 0)
+	    {
+	      fprintf (stderr, "First line in line list!");
+	      fprintf (stderr, "next_line = %f\n",
+		       storage->
+		       line_list_nu[rpacket_get_next_line_id (packet) + 1]);
+	    }
+	  else
+	    {
+	      fprintf (stderr, "last_line = %f\n",
+		       storage->
+		       line_list_nu[rpacket_get_next_line_id (packet) - 1]);
+	      fprintf (stderr, "next_line = %f\n",
+		       storage->
+		       line_list_nu[rpacket_get_next_line_id (packet) + 1]);
+	    }
+	  fprintf (stderr, "ERROR: Comoving nu less than nu_line!\n");
+	  fprintf (stderr, "comov_nu = %f\n", comov_nu);
+	  fprintf (stderr, "nu_line = %f\n", nu_line);
+	  fprintf (stderr, "(comov_nu - nu_line) / nu_line = %f\n",
+		   (comov_nu - nu_line) / nu_line);
+	  fprintf (stderr, "r = %f\n", r);
+	  fprintf (stderr, "mu = %f\n", mu);
+	  fprintf (stderr, "nu = %f\n", nu);
+	  fprintf (stderr, "doppler_factor = %f\n", doppler_factor);
+	  fprintf (stderr, "cur_zone_id = %d\n", cur_zone_id);
+	  ret_val = TARDIS_ERROR_COMOV_NU_LESS_THAN_NU_LINE;
+	}
+      else
+	{
+	  *result = ((comov_nu - nu_line) / nu) * C * t_exp;
+	}
+    }
+  return ret_val;
+}
+
+INLINE double
+compute_distance2electron (rpacket_t * packet, storage_model_t * storage)
+{
+  if (rpacket_get_virtual_packet (packet) > 0)
+    {
+      return MISS_DISTANCE;
+    }
+  double inverse_ne =
+    storage->
+    inverse_electron_densities[rpacket_get_current_shell_id (packet)] *
+    storage->inverse_sigma_thomson;
+  return rpacket_get_tau_event (packet) * inverse_ne;
+}
+
+INLINE void
+montecarlo_compute_distances (rpacket_t * packet, storage_model_t * storage)
+{
+  // Check if the last line was the same nu as the current line.
+  if (rpacket_get_close_line (packet))
+    {
+      // If so set the distance to the line to 0.0
+      rpacket_set_d_line (packet, 0.0);
+      // Reset close_line.
+      rpacket_set_close_line (packet, false);
+    }
+  else
+    {
+      rpacket_set_d_boundary (packet,
+			      compute_distance2boundary (packet, storage));
+      double d_line;
+      compute_distance2line (packet, storage, &d_line);
+      rpacket_set_d_line (packet, d_line);
+      rpacket_set_d_electron (packet,
+			      compute_distance2electron (packet, storage));
+    }
+}
