+++ tardis/montecarlo/src/search_routines.c
@@ -1,0 +1,112 @@
+#include "cmontecarlo.h"
+
+INLINE tardis_error_t
+line_search (double *nu, double nu_insert, int64_t number_of_lines,
+	     int64_t * result)
+{
+  tardis_error_t ret_val = TARDIS_ERROR_OK;
+  int64_t imin, imax;
+  imin = 0;
+  imax = number_of_lines - 1;
+  if (nu_insert > nu[imin])
+    {
+      *result = imin;
+    }
+  else if (nu_insert < nu[imax])
+    {
+      *result = imax + 1;
+    }
+  else
+    {
+      ret_val = reverse_binary_search (nu, nu_insert, imin, imax, result);
+      *result = *result + 1;
+    }
+  return ret_val;
+}
+
+inline tardis_error_t
+reverse_binary_search (double *x, double x_insert,
+		       int64_t imin, int64_t imax, int64_t * result)
+{
+  /*
+     Have in mind that *x points to a reverse sorted array.
+     That is large values will have small indices and small ones
+     will have large indices.
+   */
+  tardis_error_t ret_val = TARDIS_ERROR_OK;
+  if (x_insert > x[imin] || x_insert < x[imax])
+    {
+      ret_val = TARDIS_ERROR_BOUNDS_ERROR;
+    }
+  else
+    {
+      int imid = (imin + imax) / 2;
+      while (imax - imin > 2)
+	{
+	  if (x[imid] < x_insert)
+	    {
+	      imax = imid + 1;
+	    }
+	  else
+	    {
+	      imin = imid;
+	    }
+	  imid = (imin + imax) / 2;
+	}
+      if (imax - imin == 2 && x_insert < x[imin + 1])
+	{
+	  *result = imin + 1;
+	}
+      else
+	{
+	  *result = imin;
+	}
+    }
+  return ret_val;
+}
+
+inline tardis_error_t
+binary_search (double *x, double x_insert, int64_t imin,
+	       int64_t imax, int64_t * result)
+{
+  /*
+     Have in mind that *x points to a sorted array.
+     Like [1,2,3,4,5,...]
+   */
+  int imid;
+  tardis_error_t ret_val = TARDIS_ERROR_OK;
+  if (x_insert < x[imin] || x_insert > x[imax])
+    {
+      ret_val = TARDIS_ERROR_BOUNDS_ERROR;
+    }
+  else
+    {
+      while (imax >= imin)
+	{
+	  imid = (imin + imax) / 2;
+	  if (x[imid] == x_insert)
+	    {
+	      *result = imid;
+	      break;
+	    }
+	  else if (x[imid] < x_insert)
+	    {
+	      imin = imid + 1;
+	    }
+	  else
+	    {
+	      imax = imid - 1;
+	    }
+	}
+      if (imax - imid == 2 && x_insert < x[imin + 1])
+	{
+	  *result = imin;
+	}
+      else
+	{
+	  *result = imin;
+	}
+    }
+  return ret_val;
+}
+
